+++ factcheck/cmd/api/internal/handler/crud.go
@@ -1,4 +1,4 @@
-package handlers
+package handler
 
 import (
 	"context"
