+++ factcheck/cmd/api/internal/handler/handlers_test.go
@@ -1,7 +1,7 @@
 //go:build integration_test
 // +build integration_test
 
-package handlers_test
+package handler_test
 
 import (
 	"bytes"
